[Sandeep] Modify finding maximum of 3 floats

+++ FindMaximumProblem/Program.cs
@@ -11,15 +11,15 @@ namespace FindMaximumProblem
         static void Main(string[] args)
         {
             Console.WriteLine("Enter first value:");
-            int first = Convert.ToInt32(Console.ReadLine());
+            float first = Convert.ToSingle(Console.ReadLine());
 
             Console.WriteLine("Enter second value:");
-            int second = Convert.ToInt32(Console.ReadLine());
+            float second = Convert.ToSingle(Console.ReadLine());
 
             Console.WriteLine("Enter third value:");
-            int third = Convert.ToInt32(Console.ReadLine());
+            float third = Convert.ToSingle(Console.ReadLine());
 
-            FindMaxGeneric<int> findMax = new FindMaxGeneric<int>(first, second, third);
+            FindMaxGeneric<float> findMax = new FindMaxGeneric<float>(first, second, third);
             findMax.TestMaximum();
 
             Console.ReadLine();
